Battle classes added for both approaches
