+++ src/main/java/org/javabasictraining/openedApproach/Battle.java
@@ -1,0 +1,19 @@
+package org.javabasictraining.openedApproach;
+
+public class Battle {
+
+    private Battle() {}
+
+/** Code not used in this case of openedApproach (same one has been already used in encapsulatedApproach) */
+//    public static boolean DuelFight(Warrior attacker, Warrior defender) {
+//        while (attacker.isAlive() && defender.isAlive()) {
+//            attacker.hit(defender);
+//            if (!defender.isAlive()) {
+//                return defender.isAlive();
+//            } else {
+//                defender.hit(attacker);
+//            }
+//        }
+//        return defender.isAlive();
+//    }
+}
